Introduce max-h-65vh
I need this to refactor the featureinfobox and sub-classes.

Ideally we could add this class via tailwind.config.js but for some reason it does not work; the class does not appear in the index.css file.

As a workaround I added it differently, but still we should probably figure out why the config file does not work as expeced.