Update tree_node.json
The Spanish translation "es" was made. In this window as title, I think it is better to consider tag: species over tag: name.  Since trees generally do not have a name, unless they are considered monumental or extraordinary (as in the example of the photo). On the other hand tag:species, includes all of them. species identifies a tree better.